test: split approval ack link smoke scenario
